+++ Inventory/src/main/java/com/bartolay/inventory/services/CountryService.java
@@ -7,9 +7,7 @@ import javax.validation.Valid;
 
 import org.json.JSONObject;
 
-import com.bartolay.inventory.entity.Category;
 import com.bartolay.inventory.entity.Country;
-import com.bartolay.inventory.form.CategoryForm;
 import com.bartolay.inventory.form.CountryForm;
 
 public interface CountryService {
@@ -18,6 +16,6 @@ public interface CountryService {
 	public JSONObject retrieveDatatableList(Map<String, String> requestMap);
 	public Country create(@Valid CountryForm countryForm);
 	public Country update(@Valid CountryForm countryForm);
-	public Country delete(Long id);
+	public Country delete(int id);
 	
 }
